feat: update formatCSV to be a TS file so we can get better types later (reduces its genericity)
